requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.7
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
opencv-python-headless==4.12.0.88
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0


